Move to one-ServiceBrokerClient-per-Stage architecture
Conceptually, we want the ServiceBrokerClient to be the tool that the
Stage uses to retrieve information about other stages, so that it can
forward requests to them -- mostly things like IPs, and stuff like that.
To this extent, the SBC can be basically implicit in the way a stage is
instantiated. Perhaps we will pass configuration information to the
Stage when we spin it up, but mostly the SBC is an implementation detail
of the Stage, in that it is simply how the Stage discovers other Stages.

Right now, this is not how we have implemented it. There is currently
one SBC that has an instance of every type of Stage. This doesn't really
work if Stages are on different boxes, because then every box would have
to have an instance of every Stage.

Hence, this commit will move us towards the model of one SBC per Stage.