Make functions in notebook a bit more JavaScripty.
In JavaScript, callbacks usually appear last in a function's formal
parameters. Additionally, in classic CPS literature (for example,
Appel's _Compiling with Continuations_), the continuation also usually
appears last.

Currently, however, the experimental Continuum API in the notebook test
largely expects callbacks to appear first in the formal parameters,
including in `forward` which is our version of `callcc`.

This commit will fix this issue to make it more consistent with prior
art. We also move away from using the variable `k` and `K<T>` to refer
to continuations, as this is confusing to people who are not hardcore PL
theorists.