Remove TypeScript trnaspilation step in `Stage.forward`
Initially we had thought that (for some TypeScript function `f`) calling
`f.toString()` would print the TS source, rather than the JS source. As
a result, we would have to transpile the TS source into JS before we
`eval`d it on the `Stage` that is recieving the continuation.

As @johnshew points out, this is not true. `f.toString()` outputs JS.
Therefore the transpile step is not necessary.

This commit will remove the transpilation step from `forward`.