Add simple prototype `ServiceBrokerServer`
In a typical Continuum cluster (if you can call any Continuum cluster
"typical" given the early stage of the project) a `Stage` is running on
some node, somewhere. To talk to other nodes, it must interact with a
so-called "ServiceBroker" which will tell it where other `Stage`s are
located, kind of like a service discovery mechanism.

There are two parts to this: (1) a `ServiceBrokerClient`, which talks to
(2) a `ServiceBrokerServer`. Currently only the first of these is
implemented.

This commit will add a simple prototype version of the second.