Add simple service configurator
The `ServiceBrokerServer` currently has a bunch of services hardwired
into it, all assumed to be on localhost. This is clearly not the
eventual goal -- we want to be able to add, delete, and checkpoint
services the SBS knows about. And any way, hardwiring services into the
SBS means that it is only useful for out notebook test.

This commit will add a `ServiceConfigurator` which will be the core data
structure used to keep track of services that the SBS knows about. We
will also include the ability to serialize the services to JSON, as well
as re-hydrate them from JSON.