Move Service Broker classes to `src/core/ServiceBroker.ts`
Currently the concept of a "Service Broker" is nebulously separated
across 3 classes that are contained in `test/test-notebook.ts`. To
encourage separation of concerns and modularity, we'd like to have them
in their own module, where it will become easier to do things like
manage dependencies.

This commit will move the classes `ServiceBrokerServer`,
`ServiceBrokerClient`, and `ServiceConfigurator` into the
`src/core/ServiceBroker.ts` file to begin the process of accomplishing
these goals.