Implement stage registration on the `ServiceBrokerServer`
Currently the `ServiceBrokerServer` will configure itself manually by opening a
file containing static information about the running stages. That is, it will
open a file that contains a list of stages, their IPs, etc. This is clearly a
flawed approach, as there is no way for `Stages` to register with the SBS.

This commit will implement an endpoint on the `ServiceBrokerService` that will
allow `Stage`s to register.