Add prototype `continua broker provision` command to CLI
Currently, provisioning a service broker on a Mesos cluster is a manual
process. We'd like good tools for simply doing it at the command line.
This commit will introduce a command to the CLI that will allow us to do
this.

The command will provision a service broker on a Mesos cluster, using
Marathon. The user will provide a name that we will use locally to refer
to this service broker, which will be placed in the local broker
registry.

The command has the form `continua broker provision <marathonUrl>
<name>`, where `<marathonUrl>` is the URL of the Marathon instance that
you'd like to provision the broker at, and `<name>` is the name you'd
like to use locally to refer to the service broker. Both the `<name>`
and the `<marathonUrl>` will be recorded in the brokers registry
locally.