Add `continua broker start <port>` command
Right now, starting a service broker locally requires manually running
something like `node ./dist/src/ServiceBroker/main.js`. A better
alternative would be to have a command directly in the CLI that would
allow us to simply start a service broker locally.

This commit will introduce such a command. The form of the command is:
`continua broker start <port>`, where `<port>` is a number like 8090.
This will cause the broker to start and listen on that port.