Shorten heartbeats length to make Examples test run
Currently there is an `Examples` folder that contains a simple Continua
application. This codepath is exercised in `test/test-notebook.ts`. In
recent commits this example has stopped working; the request simply
hangs forever.

The core of the problem is that right now stages know about other stages
by periodically polling the service broker for changes to stages that
are registered and ready to receive requests. Right now, that heartbeat
is 5 seconds, and if a bunch of stages register at approximately the
same time, there is a reasonable probability they will have to poll
again before they see each other. Finally, when you run the tests, this
means that we will send a request, but the stages can't forward it
because they don't know about each other.

This commit changes the heartbeat/test cadence so that the test stages
(probably) poll the service broker before the test sends a request to be
forwarded.