Snapshot: working multi-node Continua stages.
One of the big problems in the current way of doing things is that
every stage has a `Continuum` object that needs to have instances of
other stages in order to forward to them.

As implemented, this is hard to deal with. It means that every stage
must have an instance of every other stage running. It also makes stage
discovery weird -- it is hard to add a new stage that no stage has ever
seen before, without updating the actual code of each stage and
restarting the fleet. Ick.

In the long term we want to think more about how to deploy groups of
"virtual" stages that are semantically linked, but right now we're not
quite there. There are some complications to this such as: how does
deployment work? It is likely we can borrow concepts from Orleans or
Kubernetes's pods model, but we haven't thought them through yet.

So right now, we're rolling back to the shared-nothing,
everything-is-dynamically-discovered model. We'll re-evaluate the pods
model later.